pyproject.toml
[project]
name = "voice_assistant_bot"
version = "0.1.0"
description = "Voice AI bot built with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[deepgram,elevenlabs,google,local-smart-turn-v3,runner,silero,webrtc]",
    "aiosqlite>=0.20.0,<1",
    "aiohttp-cors",
    "python-dotenv>=1.0.0,<2",
    "twilio>=9.0.0,<10"
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]